value changer renamed to counter
